Time: 20 ms (73.50%), Space: 22.3 MB (92.52%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -26,7 +26,7 @@ public:
             while (cnt--) {
                 // approach by level
                 cur = Q.front(); Q.pop();
-                curAver += (double) cur->val;
+                curAver += cur->val;
                 if (cur->left) Q.push(cur->left);
                 if (cur->right) Q.push(cur->right);
             }
